+++ lib/media_types/dsl.rb
@@ -7,14 +7,18 @@ require 'media_types/validations'
 
 module MediaTypes
   module Dsl
+
     def self.included(base)
       base.extend ClassMethods
       base.class_eval do
         class << self
+          attr_accessor :media_type_name_for, :media_type_combinations
+
           private
 
           attr_accessor :media_type_constructable, :symbol_base, :media_type_registrar, :media_type_validations
         end
+        base.media_type_combinations = Set.new
       end
     end
 
@@ -26,18 +30,26 @@ module MediaTypes
         end
       end
 
-      def valid?(output, media_type = to_constructable, **opts)
-        validations.find(String(media_type)).valid?(output, backtrace: ['.'], **opts)
+      def valid?(output, **opts)
+        to_constructable.valid?(output, **opts)
       end
 
-      def validate!(output, media_type = to_constructable, **opts)
-        validations.find(String(media_type)).validate(output, backtrace: ['.'], **opts)
+      def valid_unsafe?(output, media_type = to_constructable, **opts)
+        validations.find(media_type).valid?(output, backtrace: ['.'], **opts)
+      end
+      
+      def validate!(output, **opts)
+        to_constructable.validate!(output, **opts)
+      end
+
+      def validate_unsafe!(output, media_type = to_constructable, **opts)
+        validations.find(media_type).validate(output, backtrace: ['.'], **opts)
       end
 
       def validatable?(media_type = to_constructable)
         return false unless validations
 
-        validations.find(String(media_type), -> { nil })
+        validations.find(media_type, -> { nil })
       end
 
       def register
@@ -46,22 +58,58 @@ module MediaTypes
           registerable
         end
       end
+      
+      def view(v)
+        to_constructable.view(v)
+      end
+      def version(v)
+        to_constructable.version(v)
+      end
+      def suffix(s)
+        to_constructable.suffix(s)
+      end
+
+      def identifier_format
+        self.media_type_name_for = Proc.new do |type:, view:, version:, suffix:|
+          yield(type: type, view: view, version: version, suffix: suffix)
+        end
+      end
+
+      def identifier
+        to_constructable.to_s
+      end
+
+      def available_validations
+        self.media_type_combinations.map do |a|
+          _, view, version, suffix = a
+          view(view).version(version).suffix(suffix)
+        end
+      end
 
       private
 
-      def media_type(name, defaults: {})
+      def use_name(name, defaults: {})
+        if self.media_type_name_for.nil?
+          self.media_type_name_for = Proc.new do |type:, view:, version:, suffix:|
+            resolved_org = nil
+            if defined?(organisation)
+              resolved_org = organisation
+            else
+              resolved_org = MediaTypes::get_organisation(self)
 
-        unless defined?(:base_format)
-          define_method(:base_format) do
-            raise format('Implement the class method "base_format" in %<klass>s', klass: self)
+              raise format('Implement the class method "organisation" in %<klass>s or specify a global organisation using MediaTypes::set_organisation', klass: self) if resolved_org.nil?
+            end
+            raise ArgumentError, 'Unable to create a name for a schema with a nil name.' if type.nil?
+            raise ArgumentError, 'Unable to create a name for a schema with a nil organisation.' if resolved_org.nil?
+
+            result = "application/vnd.#{resolved_org}.#{type}"
+            result += ".v#{version}" unless version.nil?
+            result += ".#{view}" unless view.nil?
+            result += "+#{suffix}" unless suffix.nil?
+            result
           end
         end
-
-        self.media_type_constructable = Constructable.new(self, format: base_format, type: name)
-                                                     .version(defaults.fetch(:version) { nil })
-                                                     .suffix(defaults.fetch(:suffix) { nil })
-                                                     .view(defaults.fetch(:view) { nil })
-        self
+        self.media_type_constructable = Constructable.new(self, type: name).suffix(defaults.fetch(:suffix) { nil })
       end
 
       def defaults(&block)