+++ lib/media_types/scheme.rb
@@ -17,7 +17,12 @@ require 'media_types/scheme/output_empty_guard'
 require 'media_types/scheme/output_type_guard'
 require 'media_types/scheme/rules_exhausted_guard'
 
+require 'json'
+
 module MediaTypes
+  class AssertionError < StandardError
+  end
+
   ##
   # Media Type Schemes can validate content to a media type, by itself. Used by the `validations` dsl.
   #
@@ -346,6 +351,21 @@ module MediaTypes
       end.link(*args, **opts, &block)
     end
 
+    ##
+    # Mark object as a valid empty object
+    #
+    # @example Empty object
+    #
+    #   class MyMedia
+    #     include MediaTypes::Dsl
+    #
+    #     validations do
+    #       empty
+    #     end
+    #   end
+    def empty
+    end
+
     def inspect(indentation = 0)
       tabs = '  ' * indentation
       [
@@ -353,6 +373,23 @@ module MediaTypes
         rules.inspect(indentation + 1),
         "#{tabs}[/Scheme]"
       ].join("\n")
+    end
+
+    def assert_pass(fixture)
+      json = JSON.parse(fixture)
+
+      validate(json)
+    end
+    
+    def assert_fail(fixture)
+      json = JSON.parse(fixture)
+
+      begin
+        validate(json)
+      rescue MediaTypes::Scheme::ValidationError
+        return
+      end
+      raise AssertionError
     end
 
     private