+++ test/media_types/global_organisation_test.rb
@@ -1,0 +1,48 @@
+# frozen_string_literal: true
+
+require_relative '../test_helper'
+
+module MediaTypes
+  module OrganisationScope
+    def media_types_organisation
+      'longstring'
+    end
+    class GlobalOrganisationTest < Minitest::Test
+
+      class AnyType
+        include MediaTypes::Dsl
+
+
+        use_name 'test'
+
+        validations do
+          empty
+        end
+      end
+
+      def test_module_organisations
+        MediaTypes::set_organisation MediaTypes::OrganisationScope.itself, 'universal.exports'
+        assert_equal AnyType.identifier, 'application/vnd.universal.exports.test'
+      end
+
+      module Acme
+        MediaTypes::set_organisation Acme, 'acme'
+
+        class FooValidator
+          include MediaTypes::Dsl
+
+          use_name 'foo'
+
+          validations do
+            attribute :foo, String
+          end
+        end
+      end
+      
+      def test_readme_example
+        assert_equal Acme::FooValidator.identifier, 'application/vnd.acme.foo'
+      end
+
+    end
+  end
+end