+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.0.0
+ - Added the ability to do inline tests when defining validations using `assert_pass '<json>'` and `assert_fail '<json>'`.
+ - `media_type` has been replaced with `use_name`.
+ - It is no longer possible to set a default version. Please use `version <x> do` instead.
+ - You no longer need to specify a custom format string. If you set an organisation with `def self.organisation` or set a module wide organisation with `MediaTypes::set_organisation <module>, '<organisation>'` the library will generate identifiers for you.
+ - `self.base_format` has been replaced by `identifier_format do |type:, view:, version:, suffix:|`.
+ - Added the `empty` validation to mark an empty object as valid.
+ - Added the `identifier` function to get the [Media Type Identifier](https://en.wikipedia.org/wiki/Media_type) for the validator.
+ - Added `version(x)` and `view(x)` functions.
+ - Added an `available_validations` functions that returns all defined validations.
+ - Fixed an issue where validations could accidentally merge if defined with a bad `base_format`.
+ - Fixed an issue where undefined validations would accept an empty object.
+
 # 0.6.2
 
 - Fix handling empty collections